API 구조 RESTful 개선 - GitHub 레포 조회 분리
- GithubRepositoryController 신규: GET /api/v1/repositories
  * 사용자의 모든 GitHub 레포 조회 (projectId 무관)
  * 개인 소유 + 협력자 + organization 레포 포함

- RepositoryController 정리:
  * GET /api/v1/projects/{id}/repositories/github-available 제거
  * POST /api/v1/projects/{id}/repositories/sync-from-github → sync로 이름 변경
  * 프로젝트 레포만 조회하도록 정리
  * import 정리

API 엔드포인트:
- GET /api/v1/repositories: 사용자 GitHub 레포 목록
- GET /api/v1/projects/{id}/repositories: 프로젝트 레포 목록
- POST /api/v1/projects/{id}/repositories/sync: 레포 동기화
- POST/GET/DELETE /api/v1/projects/{id}/repositories: 레포 관리
