+++ data/testdb.lock.db
@@ -1,6 +1,6 @@
 #FileLock
-#Sun Apr 12 21:03:53 KST 2026
+#Sun Apr 12 21:31:48 KST 2026
 hostName=localhost
-id=19d8193b259860f9589f07d3e2d4fa2f2bad5e09eb1
+id=19d81ad422e3d021e7c6f9e52c1234f5a4e61a4e9b9
 method=file
-server=localhost\:62245
+server=localhost\:63056
